Breaking EventListActivity into EventListFragment
