+++ app/build.gradle
@@ -63,6 +63,7 @@ dependencies {
     compile 'com.android.support:design:25.2.0'
     compile 'com.android.support:support-v4:25.2.0'
     compile 'com.android.support:cardview-v7:25.2.0'
+    compile 'com.android.support:support-v13:25.2.0'
     compile 'com.android.support:recyclerview-v7:25.2.0'
     compile 'com.android.support.constraint:constraint-layout:1.0.1'
