+++ app/src/main/java/com/icebreakers/nexxus/activities/EventDetailsActivity.java
@@ -9,7 +9,6 @@ import android.support.v7.app.AppCompatActivity;
 import android.text.Html;
 import android.view.MenuItem;
 import android.view.View;
-
 import com.google.android.gms.maps.CameraUpdate;
 import com.google.android.gms.maps.CameraUpdateFactory;
 import com.google.android.gms.maps.GoogleMap;
@@ -24,7 +23,6 @@ import com.icebreakers.nexxus.databinding.ActivityEventDetailsBinding;
 import com.icebreakers.nexxus.models.MeetupEvent;
 import com.icebreakers.nexxus.models.Venue;
 import com.icebreakers.nexxus.utils.MapUtils;
-
 import org.parceler.Parcels;
 
 import java.text.SimpleDateFormat;
@@ -33,6 +31,7 @@ import java.util.Date;
 public class EventDetailsActivity extends AppCompatActivity {
 
     private static final String TAG = NexxusApplication.BASE_TAG + EventDetailsActivity.class.getName();
+    public static final String EVENT_EXTRA = "event";
     private MeetupEvent event;
 
     ActivityEventDetailsBinding binding;
@@ -70,7 +69,7 @@ public class EventDetailsActivity extends AppCompatActivity {
         getSupportActionBar().setDisplayShowHomeEnabled(true);
 
         Intent detailsIntent = getIntent();
-        event = Parcels.unwrap(detailsIntent.getParcelableExtra("event"));
+        event = Parcels.unwrap(detailsIntent.getParcelableExtra(EVENT_EXTRA));
 
         getSupportActionBar().setTitle("");
