+++ app/src/main/java/com/icebreakers/nexxus/helpers/Router.java
@@ -4,12 +4,15 @@ import android.content.Intent;
 import android.util.Log;
 import com.icebreakers.nexxus.NexxusApplication;
 import com.icebreakers.nexxus.activities.BaseActivity;
+import com.icebreakers.nexxus.activities.EventDetailsActivity;
 import com.icebreakers.nexxus.activities.EventListActivity;
 import com.icebreakers.nexxus.activities.ProfileActivity;
 import com.icebreakers.nexxus.activities.ProfileListActivity;
+import com.icebreakers.nexxus.models.MeetupEvent;
 import com.icebreakers.nexxus.models.Profile;
 import org.parceler.Parcels;
 
+import static com.icebreakers.nexxus.activities.EventDetailsActivity.EVENT_EXTRA;
 import static com.icebreakers.nexxus.activities.ProfileActivity.PROFILE_EXTRA;
 
 /**
@@ -37,5 +40,11 @@ public class Router {
         Intent intent = new Intent(activity, ProfileListActivity.class);
         intent.putExtra(PROFILE_EXTRA, Parcels.wrap(profile));
         activity.startActivity(intent);
+    }
+
+    public static void startEventDetailActivity(BaseActivity activity, MeetupEvent event) {
+        Intent detailsActivityIntent = new Intent(activity, EventDetailsActivity.class);
+        detailsActivityIntent.putExtra(EVENT_EXTRA, Parcels.wrap(event));
+        activity.startActivity(detailsActivityIntent);
     }
 }
