+++ app/src/main/java/com/icebreakers/nexxus/listeners/EventClickListener.java
@@ -1,0 +1,12 @@
+package com.icebreakers.nexxus.listeners;
+
+import com.icebreakers.nexxus.models.MeetupEvent;
+
+/**
+ * Created by amodi on 4/12/17.
+ */
+
+public interface EventClickListener {
+
+    void onEventClick(MeetupEvent meetupEvent);
+}
